requirements.txt
aiohttp==3.12.15
fastapi==0.116.1
httpx==0.28.1
openai==1.101.0
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
uvicorn==0.35.0
